pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper_translator"
version = "0.1.0"
description = "基于Whisper的音视频中文字幕生成和翻译工具"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
    { name = "legibet" }
]
keywords = ["whisper", "subtitle", "translation", "chinese"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "openai>=1.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "pydub>=0.25.1",
    "faster-whisper",
]

[project.urls]
Homepage = "https://github.com/legibet/whisper-translator-cn"
Repository = "https://github.com/legibet/whisper-translator-cn.git"

[project.scripts]
wtrans = "whisper_translator.cli:run"
wtrans-init = "whisper_translator.cli:init"

[tool.setuptools]
packages = ["whisper_translator"]

[tool.setuptools.package-data]
whisper_translator = ["config.yaml.template"]
